ver 1.0 add new setting in config

+++ WireGuard Start Stop.sh
@@ -1,14 +1,16 @@
 #!/bin/bash
 
+WG_INTERFACE_NAME="wg0-client"
+
 CLIENT=$(wg show interfaces)
 
-if [ "$CLIENT" == "wg0-client" ]; then
-	echo -e "\e[1;33mwg0-client запущен Останавливаю...\e[0m"
-	wg-quick down wg0-client
+if [ "$CLIENT" == "$WG_INTERFACE_NAME" ]; then
+	echo -e "\e[1;33m$WG_INTERFACE_NAME запущен Останавливаю...\e[0m"
+	wg-quick down $WG_INTERFACE_NAME
 	echo -e "\e[31mВыполнено.\e[0m"
-else 
-	echo -e "\e[1;33mЗапускаю wg0-client\e[0m"
-	wg-quick up wg0-client
+else
+	echo -e "\e[1;33mЗапускаю $WG_INTERFACE_NAME\e[0m"
+	wg-quick up $WG_INTERFACE_NAME
 	sudo wg show
 	echo -e "\e[31mВыполнено.\e[0m"
 	#Запускаем индикатор.
@@ -19,4 +21,7 @@ echo ""
 echo -e "\e[1;36mДля закрытия окна, нажмите любую клавишу...\e[0m"
 
 read
-rm nohup.out
+
+if [ -f nohup.out ]; then
+    rm nohup.out
+fi
